esgf-playground-utils/pyproject.toml
[tool.poetry]
name = "esgf-playground-utils"
version = "0.1.0"
description = ""
authors = ["david.poulter <david.poulter@stfc.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.3.4"
stac-pydantic = "^3.1.1"
aiokafka = "^0.11.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
isort = "^5.13.2"
ruff = "^0.5.2"
black = "^24.4.2"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
